[chore] remove duplicates in test cases
